% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_info.R
\name{bootstrap_info}
\alias{bootstrap_info}
\title{bootstrap_infor is used to estimate MSE.}
\usage{
bootstrap_info(
  dom,
  y,
  x,
  wts = NULL,
  index_d,
  Xbar,
  fit,
  useparallel = FALSE,
  ncores = 4,
  B = 200,
  conf = 0.95,
  seed = 123456
)
}
\arguments{
\item{dom}{vector for domain.}

\item{y}{response.}

\item{x}{if needs intercept, need to add before fitting the model.}

\item{wts}{sampling weights.}

\item{index_d}{the column index for different coefficients.}

\item{Xbar}{Xbar is a m * (p+1) matrix, p is the number of column of x; the first column is the dom index.}

\item{fit}{object from est_area_clut.}

\item{useparallel}{if using parallel.}

\item{ncores}{number of cores.}

\item{B}{bootstrap replicates.}

\item{conf}{confidence level.}
}
\description{
bootstrap_infor is used to estimate MSE.
}
