% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clust_nested_error.R
\name{Clust_nested_error}
\alias{Clust_nested_error}
\title{Clust_nested linear model without sampling weights}
\usage{
Clust_nested_error(
  dom,
  y,
  x,
  index_d,
  lam,
  nu = 1,
  gam = 3,
  lam0 = 0.001,
  maxiter = 500,
  tol = 0.001
)
}
\arguments{
\item{x}{x without intercept, will add intercept}
}
\description{
Clust_nested linear model without sampling weights
}
