% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clust_nested_error2.R
\name{Clust_nested_error2}
\alias{Clust_nested_error2}
\title{Clust_nested linear model without sampling weights, a different initial and warm start}
\usage{
Clust_nested_error2(
  dom,
  y,
  x,
  index_d,
  lam,
  init0 = NULL,
  vv0 = NULL,
  nu = 1,
  gam = 3,
  lam0 = 0.001,
  maxiter = 500,
  tol = 0.001
)
}
\arguments{
\item{dom}{vector for domain}

\item{y}{response}

\item{x}{if needs intercept, need to add before fitting the model}

\item{index_d}{the column index for different coef}
}
\description{
Clust_nested linear model without sampling weights, a different initial and warm start
}
