% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clust_nested_info.R
\name{Clust_nested_info}
\alias{Clust_nested_info}
\title{Clust_nested linear model  with sampling weights}
\usage{
Clust_nested_info(
  dom,
  y,
  x,
  wts,
  index_d,
  lam,
  nu = 1,
  gam = 3,
  lam0 = 0.001,
  maxiter = 500,
  tol = 0.001
)
}
\arguments{
\item{x}{x without intercept, will add intercept}

\item{wts}{is the sampling weights}
}
\description{
Clust_nested linear model  with sampling weights
}
