% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_area_clust.R
\name{est_area_clust}
\alias{est_area_clust}
\title{est_area_clust is used to estimate domain means based on unit level models}
\usage{
est_area_clust(
  dom,
  y,
  x,
  wts = NULL,
  index_d,
  Xbar,
  lamvec,
  algorithm = 2,
  init0 = NULL,
  vv0 = NULL
)
}
\arguments{
\item{dom}{vector for domain.}

\item{y}{response.}

\item{x}{if needs intercept, need to add before fitting the model.}

\item{wts}{sampling weights.}

\item{index_d}{the column index for different coefficients.}

\item{Xbar}{Xbar is a m * (p+1) matrix, p is the number of column of x; the first column is the dom index.}

\item{lamvec}{a vector of lambda values}
}
\description{
est_area_clust is used to estimate domain means based on unit level models
}
