% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_clust_nested.R
\name{fit_clust_nested}
\alias{fit_clust_nested}
\title{fit the LMM model for a given cluster structure}
\usage{
fit_clust_nested(dom, y, x, wts = NULL, index_d, cluster)
}
\arguments{
\item{dom}{vector for domain}

\item{y}{response}

\item{x}{if needs intercept, need to add before fitting the model}

\item{wts}{is the sampling weights}

\item{index_d}{the column index for different coef}
}
\description{
fit the LMM model for a given cluster structure
}
