% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initial_nested.R
\name{initial_nested}
\alias{initial_nested}
\title{initial values functions}
\usage{
initial_nested(dom, y, x, index_d, wts = NULL, lam0 = 0.001)
}
\arguments{
\item{dom}{vector for domain}

\item{y}{response}

\item{x}{if needs intercept, need to add before fitting the model}

\item{index_d}{the column index for different coef}
}
\description{
initial values functions
}
